Add IgnoreSubjectTicketForQueues config option and behaviour
This adds the ability to configure a set of queues for which the
ticket number found in the subject tag of incoming emails is ignored,
forcing new tickets to be created in the queue specified by rt-mailgate.

An example use case is a setup where RTOwner is set to an address which
routes to RT (for example to a queue named 'bounces'). In order to avoid
loops, the queue should be set up to not have any scrips which send out
mail.

Without this patch, bounces which include the subject tag of the original
message in the subject line will be directed at the existing ticket,
rather than the intended behaviour of creating a new ticket in the
bounces queue. This could cause new emails to be sent out, and could
cause a mail loop.

This will only happen if the MTA sending the bounce includes the
original subject line in the subject line of bounces. One notable example
of this behaviour is Microsoft Exchange.

A more elaborate fix would be to not ignore the ticket ID completely,
but only if the specified queue did not match the ticket's queue.
In both cases it is necessary to make this configurable on a per-queue
basis, since this breaks some assumptions about RT's normal email
workflow.